+++ Kernel/Modules/AgentCustomerInformationCenterSearch.pm
@@ -137,13 +137,11 @@ sub Run {
                 User => $CustomerLogin,
             );
 
-            if ( !( grep { $_->{Value} eq $CustomerData{UserCustomerID} } @Result ) ) {
-                push @Result,
-                    {
-                    Label => $CustomerList{$CustomerLogin},
-                    Value => $CustomerData{UserCustomerID}
-                    };
-            }
+            push @Result,
+                {
+                Label => $CustomerList{$CustomerLogin},
+                Value => $CustomerData{UserCustomerID}
+                };
             last CUSTOMERLOGIN if scalar @Result >= $MaxResults;
 
         }